+++ src/cloudData/index.js
@@ -1,6 +1,7 @@
 import firebase from "react-native-firebase";
 
 import utilities from "../utilities";
+import config from '../config';
 
 const response = {
     success: null,
@@ -10,16 +11,12 @@ const response = {
 export default class CloudData {
     static listenForData(node, callback) {
         if (__DEV__) {
-            console.log("Listening at " + (__DEV__ || config.development
-                ? "development/" + node
-                : node));
+            console.log("Listening at " + config.environment + '/' + node);
         }
 
         firebase
             .database()
-            .ref((__DEV__ || config.development
-                ? "development/" + node
-                : node))
+            .ref(config.environment + '/' + node)
             .on(
             "value",
             snapshot => {
@@ -33,17 +30,13 @@ export default class CloudData {
 
     static getData(action) {
         if (__DEV__) {
-            console.log("Dispatching get at " + (__DEV__ || config.development
-                ? "development/" + action.node
-                : action.node));
+            console.log("Dispatching get at " + config.environment + '/' + action.node);
         }
 
         return new Promise(resolve => {
             firebase
                 .database()
-                .ref((__DEV__ || config.development
-                    ? "development/" + action.node
-                    : action.node))
+                .ref(config.environment + '/' + action.node)
                 .once("value")
                 .then(snapshot => {
                     response.success = true;
@@ -60,17 +53,13 @@ export default class CloudData {
 
     static setData(action) {
         if (__DEV__) {
-            console.log("Dispatching set at " + (__DEV__ || config.development
-                ? "development/" + action.node
-                : action.node));
+            console.log("Dispatching set at " + config.environment + '/' + action.node);
         }
 
         return new Promise(resolve => {
             firebase
                 .database()
-                .ref((__DEV__ || config.development
-                    ? "development/" + action.node
-                    : action.node))
+                .ref(config.environment + '/' + action.node)
                 .set(action.data)
                 .then(() => {
                     response.success = true;
@@ -87,17 +76,13 @@ export default class CloudData {
 
     static updateData(action) {
         if (__DEV__) {
-            console.log("Dispatching update at " + (__DEV__ || config.development
-                ? "development/" + action.node
-                : action.node));
+            console.log("Dispatching update at " + config.environment + '/' + action.node);
         }
 
         return new Promise(resolve => {
             firebase
                 .database()
-                .ref((__DEV__ || config.development
-                    ? "development/" + action.node
-                    : action.node))
+                .ref(config.environment + '/' + action.node)
                 .update({
                     ...action.data,
                 })
@@ -116,17 +101,13 @@ export default class CloudData {
 
     static pushData(action) {
         if (__DEV__) {
-            console.log("Dispatching push at " + (__DEV__ || config.development
-                ? "development/" + action.node
-                : action.node));
+            console.log("Dispatching push at " + config.environment + '/' + action.node);
         }
 
         return new Promise(resolve => {
             firebase
                 .database()
-                .ref((__DEV__ || config.development
-                    ? "development/" + action.node
-                    : action.node))
+                .ref(config.environment + '/' + action.node)
                 .push({
                     ...action.data,
                 })
@@ -145,17 +126,13 @@ export default class CloudData {
 
     static deleteData(action) {
         if (__DEV__) {
-            console.log("Dispatching delete at " + (__DEV__ || config.development
-                ? "development/" + action.node
-                : action.node));
+            console.log("Dispatching delete at " + config.environment + '/' + action.node);
         }
 
         return new Promise(resolve => {
             firebase
                 .database()
-                .ref((__DEV__ || config.development
-                    ? "development/" + action.node
-                    : action.node))
+                .ref(config.environment + '/' + action.node)
                 .set(null)
                 .then(() => {
                     response.success = true;